pyproject.toml
[tool.poetry]
name = "codemend"
version = "0.2.3"
description = ""
authors = ["Accord33"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


langchain = "^0.2.8"
google-generativeai = "^0.7.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
codemend = "codemend.main:main"
